% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isotope.R
\name{isotope}
\alias{isotope}
\title{<Add Title>}
\usage{
isotope(d, layoutMode = "masonry", filterCols = NULL, sortCols = NULL,
  elemTpl = NULL, lang = "en", width = NULL, height = NULL)
}
\description{
<Add Description>
}

