% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isotope.R
\name{isotopeOutput}
\alias{isotopeOutput}
\title{Widget output function for use in Shiny}
\usage{
isotopeOutput(outputId, width = "100\%", height = "600px")
}
\description{
Widget output function for use in Shiny
}

