% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isotope.R
\name{isotope}
\alias{isotope}
\title{<Add Title>}
\usage{
isotope(d, filterCols = NULL, sortCols = NULL, elemTpl = NULL,
  filterTitle = "Filter", sortTitle = "Sort", width = NULL,
  height = NULL)
}
\description{
<Add Description>
}

